refactor: Move "plugins" into "editor" dir
