+++ src/editor/index.ts
@@ -1,0 +1,1 @@
+export { createPlugins } from './plugins'
