+++ src/editor/plugins/equations/editor.tsx
