+++ src/editor/plugins/equations/index.ts
