+++ src/editor/plugins/equations/renderer.tsx
