+++ src/editor/plugins/equations/sign.tsx
