+++ src/editor/plugins/index.ts
