+++ src/editor/plugins/registry.ts
