+++ src/editor/plugins/serlo-table/index.tsx
