+++ src/editor/plugins/serlo-table/renderer.tsx
