+++ src/pages/index.tsx
@@ -4,7 +4,7 @@ import dynamic from 'next/dynamic'
 import { kitchenSink } from '../kitchen-sink'
 import { Layout } from '../components'
 import { migrate, emptyDocument } from '../storage-format'
-import { createPlugins } from '../plugins'
+import { createPlugins } from '../editor'
 import { getJsonBody } from '../utils/get-json-body'
 import { Renderer } from '@edtr-io/renderer'
 import type { EditorProps } from '../editor/editor'
